refactor: rework component uptime
