+++ frontend/package.json
@@ -17,6 +17,7 @@
     "@tanstack/react-query": "^5.59.16",
     "@tanstack/react-query-devtools": "^5.59.16",
     "@types/node": "^20.17.3",
+    "d3": "^7.9.0",
     "date-fns": "^3.6.0",
     "date-fns-tz": "^3.2.0",
     "formik": "^2.4.6",
@@ -38,6 +39,7 @@
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
+    "@types/d3": "^7.4.3",
     "@types/lodash": "^4.17.12",
     "@types/qs": "^6.9.16",
     "@types/react": "^18.3.12",
