+++ frontend/src/pages/StatusPages/ShowStatusPageIncident.tsx
@@ -11,8 +11,8 @@ import useApiService from '@/hooks/useApi'
 import { ComponentStatus } from '@/types/enums'
 import { ModelID } from '@/types/models'
 
+import ComponentStatusTimeline from './components/ComponentStatusTimeline'
 import StatusPageIncidentUpdateRow from './components/IncidentUpdateRow'
-import UptimeVisualization from './components/UptimeVis'
 
 import CreateStatusPageIncidentUpdateModal from './CreateStatusPageIncidentUpdateModal'
 import { mapComponentStatusToStyleProps, statusToTitleCase } from './utils'
@@ -133,7 +133,7 @@ const ShowStatusPageIncident = () => {
                   <h3>Uptime</h3>
                 </ContentHeader>
                 <ContentSection>
-                  {getIncidentEvents.data && <UptimeVisualization events={getIncidentEvents.data.items} />}
+                  {getIncidentEvents.data && <ComponentStatusTimeline events={getIncidentEvents.data.items} />}
                 </ContentSection>
               </ContentMain>
               <ContentSidebar>
